Add collector based example
