+++ example/metric/README.md
@@ -22,7 +22,11 @@ Next, set this project ID to the environment variable `GOOGLE_PROJECT_ID` using 
 export GOOGLE_PROJECT_ID=$(gcloud config get-value project)
 ```
 
-Once you ensure the API is enabled, then build the example application and run the executable.
+Once you ensure the API is enabled, then build the example application and run the executable. There are two separate examples showcasing the following - 
+1. Exporting metrics via the SDK.
+2. Exporting metrics via the OpenTelemetry Collector.
+
+Change the current directory to the example you wish to run - either to `sdk` directory or `collector` directory and then run the example using following commands:
 
 ```
 $ go build -o metrics
@@ -35,6 +39,8 @@ $ ./metrics
 ```
 
 To ensure you have enough data to create interesting charts for the Counter and Observer metrics generated by the example, keep the application running for at least five minutes.
+
+*Note: For running the collector example, you need to have a locally running OpenTelemetry Collector, configured using the provided [sample config](./collector/sample-collector-config.yaml). Instructions for running OpenTelemetry Collector on your system can be found [here](https://opentelemetry.io/docs/collector/getting-started/#local).*
 
 ## Create dashboard
 
@@ -53,5 +59,7 @@ $ ./create_dashboard.sh
 This script creates a dashboard titled "OpenTelemetry exporter example/metric".
 
 You should be able to view line charts like below once you create the dashboard.
+
+*Note: This script is configured to create dashboard which displays the metrics generated via the `sdk` example.*
 
 <img width="1200" alt="2 charts in dashboard" src="images/charts.png?raw=true"/>
